Fix native-width JSystem counters and heap blocks

+++ src/static/JSystem/JFramework/JFWDisplay.cpp
@@ -328,8 +328,8 @@ void waitForTick(u32 p1, u16 p2) {
             if (!OSReceiveMessage(JUTVideo::getManager()->getMessageQueue(), &msg, OS_MESSAGE_BLOCK)) {
                 msg = 0;
             }
-        } while (((int)msg - (int)nextCount) < 0);
-        nextCount = (int)msg + uVar1;
+        } while ((s32)((uintptr_t)msg - nextCount) < 0);
+        nextCount = (u32)(uintptr_t)msg + uVar1;
     }
 }
